[Feature] Redesigned the `Disk.Config`
Previously, the API on `Disk.Config` provided a method for adding
sections, that accepted `Probability`-s. It has been decided /for
testing reasons/ to pass `Optional<Probability>`-s to it.

Those are stored in a _dirty_ map of configs /which was previously
thought to be the clean final one/.

We did this since we have not thought of use cases for a clean config
map yet. We'll work around them, when we do.